Refactoring: class GeoDataPointmap::EditPointsCommand moved to independent files.
